+++ src/main/java/com/github/olegushak/FTT/client/AirportsClient.java
@@ -1,12 +1,12 @@
 package com.github.olegushak.FTT.client;
 
 import com.github.olegushak.FTT.dto.AirportDto;
-import com.mashape.unirest.http.exceptions.UnirestException;
+
 
 import java.io.IOException;
-import java.util.List;
+import java.util.Map;
 
 public interface AirportsClient {
-    List<AirportDto> retrieveAirports ()
-            throws IOException, UnirestException;
+    Map<String, AirportDto> retrieveAirports ()
+            throws IOException;
 }