+++ src/main/java/com/github/olegushak/FTT/client/AirportsClientImpl.java
@@ -1,44 +1,35 @@
 package com.github.olegushak.FTT.client;
 
-import com.fasterxml.jackson.core.type.TypeReference;
 import com.github.olegushak.FTT.dto.AirportDto;
-import com.github.olegushak.FTT.service.UniRestService;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.mashape.unirest.http.HttpResponse;
-import com.mashape.unirest.http.JsonNode;
-import org.apache.http.HttpStatus;
+import com.github.olegushak.FTT.dto.ResponseDto;
+import kong.unirest.UnirestInstance;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import java.io.IOException;
-import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
 
-import static com.github.olegushak.FTT.service.UniRestServiceImpl.AIRPORTS_FORMAT;
-import static com.github.olegushak.FTT.service.UniRestServiceImpl.JSON_KEY;
 
 @Component
 public class AirportsClientImpl implements AirportsClient {
 
-    private final UniRestService uniRestService;
-
-    private final ObjectMapper objectMapper;
+    private final UnirestInstance unirest;
+    public static final String AIRPORTS_FORMAT = "/flights/airports";
+    private static final String HOST = "https://sky-scanner3.p.rapidapi.com";
 
     @Autowired
-    public AirportsClientImpl(UniRestService uniRestService, ObjectMapper objectMapper){
-        this.uniRestService = uniRestService;
-        this.objectMapper = objectMapper;
+    public AirportsClientImpl(UnirestInstance unirest){
+        this.unirest = unirest;
     }
+
     @Override
-    public List<AirportDto> retrieveAirports() throws IOException{
+    public Map<String, AirportDto> retrieveAirports()  {
+        ResponseDto response = unirest.get(HOST + AIRPORTS_FORMAT)
+                .asObject(ResponseDto.class)
+                .getBody();
 
-        HttpResponse<JsonNode> response = uniRestService.get(AIRPORTS_FORMAT);
-        if (response.getStatus() != HttpStatus.SC_OK) {
-            return null;
-        }
-
-        String jsonList = response.getBody().getObject().get(JSON_KEY).toString();
-
-        return objectMapper.readValue(jsonList, new TypeReference<>() {
-        });
+        return response.getData().stream()
+                .map(airport -> (AirportDto) airport).filter(airport -> airport.getSkyId()!= null)
+                .collect(Collectors.toMap(AirportDto::getName, airport -> airport));
     }
 }