+++ src/main/java/com/github/olegushak/FTT/client/LocalisationClient.java
@@ -3,9 +3,9 @@ package com.github.olegushak.FTT.client;
 import com.github.olegushak.FTT.dto.LocalisationDto;
 
 import java.io.IOException;
-import java.util.List;
+import java.util.Map;
 
 public interface LocalisationClient {
 
-    List<LocalisationDto>  retrieveLocalisations() throws IOException;
+    Map<String, LocalisationDto> retrieveLocalisations() throws IOException;
 }