+++ src/main/java/com/github/olegushak/FTT/client/LocalisationClientImpl.java
@@ -1,47 +1,33 @@
 package com.github.olegushak.FTT.client;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.olegushak.FTT.dto.LocalisationDto;
-import com.github.olegushak.FTT.service.UniRestService;
-import com.mashape.unirest.http.HttpResponse;
-import com.mashape.unirest.http.JsonNode;
-import org.apache.http.HttpStatus;
+import com.github.olegushak.FTT.dto.ResponseDto;
+import kong.unirest.UnirestInstance;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-import com.fasterxml.jackson.core.type.TypeReference;
-import java.io.IOException;
-import java.util.List;
-
-import static com.github.olegushak.FTT.service.UniRestServiceImpl.JSON_KEY;
-import static com.github.olegushak.FTT.service.UniRestServiceImpl.LOCALISATION_FORMAT;
-
+import java.util.Map;
+import java.util.stream.Collectors;
 
 @Component
 public class LocalisationClientImpl implements LocalisationClient{
 
-    private final UniRestService uniRestService;
-
-
-    private final ObjectMapper objectMapper ;
+    private final UnirestInstance unirest;
+    private final String LOCALISATION_FORMAT = "/get-config";
+    public final String HOST = "https://sky-scanner3.p.rapidapi.com";
 
     @Autowired
-    public LocalisationClientImpl(UniRestService uniRestService, ObjectMapper objectMapper) {
-        this.uniRestService = uniRestService;
-        this.objectMapper = objectMapper;
+    public LocalisationClientImpl(UnirestInstance unirest) {
+        this.unirest = unirest;
     }
 
-
     @Override
-    public List<LocalisationDto> retrieveLocalisations() throws IOException{
+    public Map<String,LocalisationDto> retrieveLocalisations(){
+        ResponseDto response = unirest.get(HOST + LOCALISATION_FORMAT)
+                .asObject(ResponseDto.class)
+                .getBody();
 
-       HttpResponse<JsonNode> response = uniRestService.get(LOCALISATION_FORMAT);
-        if (response.getStatus() != HttpStatus.SC_OK) {
-            return null;
-        }
-
-        String jsonList = response.getBody().getObject().get(JSON_KEY).toString();
-
-       return objectMapper.readValue(jsonList, new TypeReference<>() {
-       });
+        return response.getData().stream()
+                .map(localisation -> (LocalisationDto)localisation )
+                .collect(Collectors.toMap(LocalisationDto::getCountry, localisation -> localisation));
     }
 }