+++ src/main/java/com/github/olegushak/FTT/dto/CarrierDto.java
@@ -2,12 +2,14 @@ package com.github.olegushak.FTT.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
 import lombok.Data;
 
 import java.util.List;
 
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
+@AllArgsConstructor
 public class CarrierDto {
 
     @JsonProperty("marketing")