+++ src/main/java/com/github/olegushak/FTT/dto/DetailsResponseDto.java
@@ -2,12 +2,14 @@ package com.github.olegushak.FTT.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
-import lombok.ToString;
 
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
-@ToString(callSuper = true)
-public class FlightDetailsDto {
+public class DetailsResponseDto {
 
-    private DetailedItineraryDto itinerary;
+    private FlightDetailsDto data;
+
+    private boolean status;
+
+    private String message;
 }