+++ src/main/java/com/github/olegushak/FTT/dto/FlightDto.java
@@ -3,12 +3,14 @@ package com.github.olegushak.FTT.dto;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
+import lombok.ToString;
 
 import java.util.List;
 
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class FlightsDto {
+@ToString(callSuper = true)
+public class FlightDto {
 
     @JsonProperty("itineraries")
     private List<ItineraryDto> itineraries;