+++ src/main/java/com/github/olegushak/FTT/dto/LegDto.java
@@ -2,14 +2,15 @@ package com.github.olegushak.FTT.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
+@AllArgsConstructor
+@NoArgsConstructor
 public class LegDto {
-
-    @JsonProperty("id")
-    private String id;
 
     @JsonProperty("origin")
     private OriginDto origin;