+++ src/main/java/com/github/olegushak/FTT/dto/ResponseDto.java
@@ -2,12 +2,17 @@ package com.github.olegushak.FTT.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
-import lombok.ToString;
+
+import java.util.List;
+
 
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
-@ToString(callSuper = true)
-public class FlightDetailsDto {
+public class ResponseDto {
 
-    private DetailedItineraryDto itinerary;
+    private List<?> data;
+
+    private boolean status;
+
+    private String message;
 }