+++ src/main/java/com/github/olegushak/FTT/service/UniRestService.java
@@ -1,9 +1,0 @@
-package com.github.olegushak.FTT.service;
-
-import com.mashape.unirest.http.HttpResponse;
-import com.mashape.unirest.http.JsonNode;
-
-public interface UniRestService {
-
-    HttpResponse<JsonNode> get(String path);
-}