+++ src/test/java/com/github/olegushak/FTT/client/AirportsClientTest.java
@@ -1,7 +1,7 @@
 package com.github.olegushak.FTT.client;
 
 import com.github.olegushak.FTT.dto.AirportDto;
-import com.mashape.unirest.http.exceptions.UnirestException;
+import kong.unirest.UnirestException;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -10,7 +10,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 
 import java.io.IOException;
-import java.util.List;
+import java.util.Map;
 
 @DisplayName("Integration-level testing for FttAirportsClientImplTest")
 @ActiveProfiles("test")
@@ -22,9 +22,9 @@ public class AirportsClientTest {
 
     @Test
     public void shouldRetrieveAirports(){
-        List<AirportDto> airportsList = null;
+        Map<String,AirportDto> airports = null;
         try {
-            airportsList = airportsClient.retrieveAirports();
+            airports = airportsClient.retrieveAirports();
         } catch (IOException e){
             e.printStackTrace();
         } catch (UnirestException e) {
@@ -32,7 +32,7 @@ public class AirportsClientTest {
         }
 
         //then
-        Assertions.assertNotNull(airportsList);
-        Assertions.assertFalse(airportsList.isEmpty());
+        Assertions.assertNotNull(airports);
+        Assertions.assertFalse(airports.isEmpty());
     }
 }