+++ src/test/java/com/github/olegushak/FTT/client/FlightsClientTest.java
@@ -1,7 +1,8 @@
 package com.github.olegushak.FTT.client;
 
 import com.github.olegushak.FTT.dto.FlightDetailsDto;
-import com.github.olegushak.FTT.dto.FlightsDto;
+import com.github.olegushak.FTT.dto.FlightDto;
+import com.github.olegushak.FTT.dto.FlightRequestArgs;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -21,9 +22,15 @@ public class FlightsClientTest {
 
     @Test
     public void shouldRetrieveOneWayFlights(){
-        FlightsDto flights = null;
+        FlightRequestArgs flightRequestArgs = FlightRequestArgs.builder()
+                .fromEntityId("HAN")
+                .toEntityId("SGN")
+                .departDate("2024-07-17")
+                .build();
+
+        FlightDto flights = null;
         try {
-            flights = flightsClient.retrieveOneWayFlights("HAN","SGN","2024-07-17");
+            flights = flightsClient.retrieveOneWayFlights(flightRequestArgs);
         } catch (IOException e){
             e.printStackTrace();
         }
@@ -36,24 +43,36 @@ public class FlightsClientTest {
 
     @Test
     public void shouldRetrieveRoundTripFlights(){
-        FlightsDto flights = null;
+        FlightRequestArgs flightRequestArgs = FlightRequestArgs.builder()
+                .fromEntityId("HAN")
+                .toEntityId("SGN")
+                .departDate("2024-07-17")
+                .returnDate("2024-07-20")
+                .build();
+
+        FlightDto flight = null;
         try {
-            flights = flightsClient.retrieveRoundTripFlights("HAN","SGN","2024-07-17","2024-07-20");
+            flight = flightsClient.retrieveRoundTripFlights(flightRequestArgs);
         } catch (IOException e){
             e.printStackTrace();
         }
 
         //then
-        Assertions.assertNotNull(flights);
-        Assertions.assertFalse(flights.getItineraries().isEmpty());
+        Assertions.assertNotNull(flight);
+        Assertions.assertFalse(flight.getItineraries().isEmpty());
 
     }
 
     @Test
     public void shouldRetrieveFlightDetails(){
+        FlightRequestArgs flightRequestArgs = FlightRequestArgs.builder()
+                .itineraryId("16240-2406212045--31703-0-12071-2406212250")
+                .token("eyJhIjoxLCJjIjowLCJpIjowLCJjYyI6ImVjb25vbXkiLCJvIjoiU0dOIiwiZCI6IkhBTiIsImQxIjoiMjAyNC0wNi0yMSJ9")
+                .build();
+
         FlightDetailsDto details = null;
         try {
-            details = flightsClient.retrieveFlightDetails("eyJhIjoxLCJjIjowLCJpIjowLCJjYyI6ImVjb25vbXkiLCJvIjoiU0dOIiwiZCI6IkhBTiIsImQxIjoiMjAyNC0wNi0yMSJ9","16240-2406212045--31703-0-12071-2406212250");
+            details = flightsClient.retrieveFlightDetails(flightRequestArgs);
         } catch (IOException e){
             e.printStackTrace();
         }