+++ src/test/java/com/github/olegushak/FTT/repository/TelegramUserRepositoryIT.java
@@ -1,5 +1,6 @@
 package com.github.olegushak.FTT.repository;
 
+import com.github.olegushak.FTT.repository.entity.Flight;
 import com.github.olegushak.FTT.repository.entity.TelegramUser;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -46,5 +47,17 @@ public class TelegramUserRepositoryIT {
 
     }
 
+    @Sql(scripts = {"/sql/clearDbs.sql", "/sql/twoFlightsForUser.sql"})
+    @Test
+    public void shouldProperlyGetAllFlightsForUser(){
+        Optional<TelegramUser> userFromDB = telegramUserRepository.findById("123456789");
 
+        Assertions.assertTrue(userFromDB.isPresent());
+        List<Flight> flights = userFromDB.get().getFlights();
+
+        Assertions.assertEquals("HAN", flights.get(0).getDeparture());
+       Assertions.assertEquals("KJA", flights.get(1).getDestination());
+        Assertions.assertEquals("15083-2402151840--30858,-31825-1-14355-2402160020", flights.get(0).getId());
+        Assertions.assertEquals("15083-2402191150--31058,-31821-1-14666-2402160908", flights.get(1).getId());
+    }
 }